Run passing tests with the GIL disabled
In `--disable-gil` builds, `make test` and `make buildbottest` now run a subset
of tests with `-Xgil=0` (tests that are not currently known to fail). The test
runner itself occasionally crashes with the GIL off, so until that's resolved,
it takes a `--disable-gil` option to pass `-Xgil=0` to the worker processes.